perf(ui): keep profile avatar URLs out of startup
