+++ ui/src/app/browser.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, it } from "vitest";
 import { CONTROL_UI_BASE_PATH_ATTRIBUTE } from "../../../src/gateway/control-ui-contract.js";
-import { resolveControlUiBasePath, resolveControlUiResourceBasePath } from "./browser.ts";
+import { resolveControlUiPaths } from "./browser.ts";
 
 afterEach(() => {
   document.documentElement.removeAttribute(CONTROL_UI_BASE_PATH_ATTRIBUTE);
@@ -10,19 +10,16 @@ describe("Control UI route and resource bases", () => {
   it("keeps a root resource mount separate from an inferred route namespace", () => {
     document.documentElement.setAttribute(CONTROL_UI_BASE_PATH_ATTRIBUTE, "");
 
-    expect(resolveControlUiBasePath("/__openclaw__/new")).toBe("/__openclaw__");
-    expect(resolveControlUiResourceBasePath("/__openclaw__/new")).toBe("");
+    expect(resolveControlUiPaths("/__openclaw__/new")).toEqual(["/__openclaw__", ""]);
   });
 
   it("uses a configured Gateway mount for both routes and resources", () => {
     document.documentElement.setAttribute(CONTROL_UI_BASE_PATH_ATTRIBUTE, "/openclaw");
 
-    expect(resolveControlUiBasePath("/openclaw/new")).toBe("/openclaw");
-    expect(resolveControlUiResourceBasePath("/openclaw/new")).toBe("/openclaw");
+    expect(resolveControlUiPaths("/openclaw/new")).toEqual(["/openclaw", "/openclaw"]);
   });
 
   it("retains pathname inference when no Gateway mount is declared", () => {
-    expect(resolveControlUiBasePath("/portable/new")).toBe("/portable");
-    expect(resolveControlUiResourceBasePath("/portable/new")).toBe("/portable");
+    expect(resolveControlUiPaths("/portable/new")).toEqual(["/portable", "/portable"]);
   });
 });
