+++ ui/src/app/user-profile.test.ts
@@ -4,7 +4,6 @@ import {
   readPresenceEntries,
   resolveCurrentSelfUser,
   resolveSelfPresenceUser,
-  userProfileAvatarUrl,
 } from "./user-profile.ts";
 
 describe("connection user profile helpers", () => {
@@ -43,27 +42,9 @@ describe("connection user profile helpers", () => {
     ).toEqual({ id: "profile-1", name: "Ada" });
   });
 
-  it("reads presence payloads and builds scoped cache-busted avatar URLs", () => {
+  it("reads presence payloads", () => {
     const entries = [{ instanceId: "self", user: { id: "profile/1" }, ts: 1 }];
     expect(readPresenceEntries({ presence: entries })).toEqual(entries);
     expect(readPresenceEntries({ presence: null })).toBeUndefined();
-    expect(
-      userProfileAvatarUrl(
-        "wss://gateway.example.test/control",
-        "profile/1",
-        42,
-        "/control",
-        "https://gateway.example.test/control/profile",
-      ),
-    ).toBe("https://gateway.example.test/control/api/users/profile%2F1/avatar?v=42");
-    expect(
-      userProfileAvatarUrl(
-        "wss://remote.example.test",
-        "profile-1",
-        "content-hash-png",
-        "",
-        "https://gateway.example.test/control/profile",
-      ),
-    ).toBe("https://remote.example.test/api/users/profile-1/avatar?v=content-hash-png");
   });
 });
