+++ ui/src/app/user-profile.ts
@@ -1,4 +1,3 @@
-import { normalizeRouteBasePath } from "@openclaw/uirouter";
 import type { PresenceEntry } from "../api/types.ts";
 
 export type AuthenticatedUser = NonNullable<PresenceEntry["user"]>;
@@ -41,44 +40,4 @@ export function resolveCurrentSelfUser({
   return snapshotUser && (!presenceUser || snapshotUser.id === presenceUser.id)
     ? snapshotUser
     : presenceUser;
-}
-
-export function userProfileAvatarUrl(
-  gatewayUrl: string,
-  profileId: string,
-  revision: string | number,
-  resourceBasePath = "",
-  documentHref?: string,
-): string | null {
-  const pageHref = documentHref ?? globalThis.location?.href;
-  if (!pageHref) {
-    return null;
-  }
-  try {
-    const url = new URL(gatewayUrl, pageHref);
-    if (url.protocol === "ws:") {
-      url.protocol = "http:";
-    } else if (url.protocol === "wss:") {
-      url.protocol = "https:";
-    }
-    // The shared avatar loader authenticates cross-origin Gateway requests and
-    // turns their response into a local blob accepted by the Control UI CSP.
-    if (!["http:", "https:"].includes(url.protocol)) {
-      return null;
-    }
-    url.username = "";
-    url.password = "";
-    const documentOrigin =
-      documentHref === undefined && globalThis.location?.origin
-        ? globalThis.location.origin
-        : new URL(pageHref).origin;
-    const sameOriginResourceBase =
-      url.origin === documentOrigin ? normalizeRouteBasePath(resourceBasePath) : "";
-    url.pathname = `${sameOriginResourceBase}/api/users/${encodeURIComponent(profileId)}/avatar`;
-    url.search = `?v=${revision}`;
-    url.hash = "";
-    return url.href;
-  } catch {
-    return null;
-  }
 }
